+++ pmd-scala-modules/pmd-scala-common/src/main/java/net/sourceforge/pmd/lang/scala/ScalaLanguageModule.java
@@ -4,17 +4,11 @@
 
 package net.sourceforge.pmd.lang.scala;
 
-import org.checkerframework.checker.nullness.qual.NonNull;
-
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.cpd.ScalaTokenizer;
 import net.sourceforge.pmd.cpd.Tokenizer;
 import net.sourceforge.pmd.lang.LanguagePropertyBundle;
 import net.sourceforge.pmd.lang.LanguageRegistry;
-import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.impl.SimpleLanguageModuleBase;
-
-import scala.meta.Dialect;
 
 /**
  * Language Module for Scala.
@@ -37,18 +31,6 @@ public class ScalaLanguageModule extends SimpleLanguageModuleBase {
 
     public static ScalaLanguageModule getInstance() {
         return (ScalaLanguageModule) LanguageRegistry.PMD.getLanguageById(ID);
-    }
-
-    @InternalApi
-    public static @NonNull Dialect dialectOf(LanguageVersion v) {
-        switch (v.getVersion()) {
-        case "2.10": return scala.meta.dialects.package$.MODULE$.Scala210();
-        case "2.11": return scala.meta.dialects.package$.MODULE$.Scala211();
-        case "2.12": return scala.meta.dialects.package$.MODULE$.Scala212();
-        case "2.13": return scala.meta.dialects.package$.MODULE$.Scala213();
-        default:
-            throw new IllegalArgumentException(v.getVersion());
-        }
     }
 
     @Override